Fix bomb-plane collision. Make lives=1 to start right now.

+++ bomb.cpp
@@ -24,11 +24,11 @@ class Bomb : public Displayable {
 			x = x + speedX;
 		}
 
-		int getX() {
+		double getX() {
 			return x;
 		}
 
-		int getY() {
+		double getY() {
 			return y;
 		}
